+++ src/ControlPanel/ControlPanel.BusinessLogic.Implementations.Unit.Tests/ControlPanel.BusinessLogic.Implementations.Unit.Tests.csproj
@@ -5,9 +5,9 @@
         <CodeAnalysisRuleSet>../../SolutionItems/stylecop.ruleset</CodeAnalysisRuleSet>
         <IsPackable>false</IsPackable>
 
-        <AssemblyName>DigitalLibrary.ControlPanel.BusinessLogic.Implementations.Tests</AssemblyName>
+        <AssemblyName>DigitalLibrary.ControlPanel.BusinessLogic.Implementations.Unit.Tests</AssemblyName>
 
-        <RootNamespace>DigitalLibrary.ControlPanel.BusinessLogic.Implementations.Tests</RootNamespace>
+        <RootNamespace>DigitalLibrary.ControlPanel.BusinessLogic.Implementations.Unit.Tests</RootNamespace>
 
         <OutputType>Library</OutputType>
