+++ src/ControlPanel/ControlPanel.BusinessLogic.Implementations.Unit.Tests/Menu/MenuBusinessLogic_DeleteAsync_Should.cs
@@ -1,25 +1,42 @@
-namespace DigitalLibrary.ControlPanel.BusinessLogic.Implementations.Tests.Menu
+namespace DigitalLibrary.ControlPanel.BusinessLogic.Implementations.Unit.Tests.Menu
 {
+    using System;
     using System.Collections.Generic;
     using System.Reflection;
     using System.Threading.Tasks;
+
+    using Exceptions.Menu;
 
     using FluentAssertions;
 
     using Xunit;
 
     [Collection(nameof(AssemblyName.GetAssemblyName))]
-    public class MenuBusinessLogic_GetAllAsync_Should : TestBase
+    public class MenuBusinessLogic_DeleteAsync_Should : TestBase
     {
-        private const string TestInfo = nameof(MenuBusinessLogic_GetAllAsync_Should);
+        private const string TestInfo = nameof(MenuBusinessLogic_DeleteAsync_Should);
 
-        public MenuBusinessLogic_GetAllAsync_Should() : base(TestInfo)
+        public MenuBusinessLogic_DeleteAsync_Should() : base(TestInfo)
         {
         }
 
         [Fact]
         [Trait("Category", "Unit")]
-        public async Task Return_AllItems()
+        public async Task Throw_DeleteAsyncOperationException_WhenInputIsNull()
+        {
+            // Arrange
+
+            // Act
+            Func<Task> action = async () => { await MenuBusinessLogic.DeleteAsync(null).ConfigureAwait(false); };
+
+            // Assert
+            action.Should().ThrowExactly<MenuBusinessLogicDeleteAsyncOperationException>()
+               .WithInnerException<MenuNullInputException>();
+        }
+
+        [Fact]
+        [Trait("Category", "Unit")]
+        public async Task Delete_AnItem()
         {
             // Arrange
             DomainModel.Entities.Module module = new DomainModel.Entities.Module
@@ -51,10 +68,11 @@ namespace DigitalLibrary.ControlPanel.BusinessLogic.Implementations.Tests.Menu
                 await MenuBusinessLogic.AddAsync(menuInactive).ConfigureAwait(false);
 
             // Act
+            await MenuBusinessLogic.DeleteAsync(menuInactiveResult).ConfigureAwait(false);
             List<DomainModel.Entities.Menu> result = await MenuBusinessLogic.GetAllAsync().ConfigureAwait(false);
 
             // Assert
-            result.Count.Should().Be(2);
+            result.Count.Should().Be(1);
         }
     }
 }
